changed minishell.c to main.c
