+++ srcs/main.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minishell.c                                        :+:      :+:    :+:   */
+/*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: diogmart <diogmart@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/11 12:07:31 by diogmart          #+#    #+#             */
-/*   Updated: 2023/04/14 11:11:29 by diogmart         ###   ########.fr       */
+/*   Updated: 2023/04/14 12:20:16 by diogmart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,18 +16,6 @@
 TODO:
 	- a lot;
 */
-
-void	ft_printlist(t_list *list)
-{
-	t_list *tmp;
-
-	tmp = list;
-	while (tmp != NULL)
-	{
-		printf("%s\n", ((t_env *)(tmp->content))->info);
-		tmp = tmp->next;
-	}
-}
 
 int	main(int argc, char **argv, char **sys_env)
 {
