+++ srcs/utils.c
@@ -6,11 +6,11 @@
 /*   By: diogmart <diogmart@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/14 10:59:10 by diogmart          #+#    #+#             */
-/*   Updated: 2023/04/14 11:09:03 by diogmart         ###   ########.fr       */
+/*   Updated: 2023/04/14 12:02:46 by diogmart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../includes/minishell.h"
 
 void	add_to_env(t_list **env, char *info)
 {
@@ -18,4 +18,16 @@ void	add_to_env(t_list **env, char *info)
 
 	new = ft_lstnew(ft_create_data(info));
 	ft_lstadd_back(env, new);
+}
+
+void	ft_printlist(t_list *list)
+{
+	t_list *tmp;
+
+	tmp = list;
+	while (tmp != NULL)
+	{
+		printf("%s\n", ((t_env *)(tmp->content))->info);
+		tmp = tmp->next;
+	}
 }
